requirements.txt
alabaster==0.7.13
allennlp==0.8.4
amqp==5.3.1
anyio==3.7.1
asgiref==3.7.2
awscli==1.31.13
Babel==2.14.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
blis==0.2.4
boto3==1.33.13
botocore==1.33.13
cached-property==1.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.4
conllu==0.11
cycler==0.11.0
cymem==2.0.11
Django==3.2.25
django-ninja==0.12.3
django_celery_results==2.6.0
djangorestframework==3.15.1
docutils==0.16
editdistance==0.6.2
exceptiongroup==1.3.0
filelock==3.12.2
flaky==3.8.1
Flask==2.2.5
Flask-Cors==5.0.0
fonttools==4.38.0
fsspec==2023.1.0
ftfy==6.1.1
gevent==22.10.2
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
h11==0.14.0
h5py==3.8.0
httptools==0.6.0
huggingface-hub==0.16.4
idna==3.10
imagesize==1.4.1
importlib-metadata==4.13.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.3.2
jsonnet==0.20.0
jsonpickle==3.4.2
kafka-python==2.2.14
kiwisolver==1.4.5
kombu==5.2.4
MarkupSafe==2.1.5
matplotlib==3.5.3
murmurhash==1.0.13
nltk==3.8.1
numpy==1.19.5
numpydoc==1.5.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
overrides==4.1.2
packaging==24.0
parsimonious==0.10.0
Pillow==9.5.0
plac==0.9.6
pluggy==1.2.0
preshed==2.0.1
promise==2.3
prompt_toolkit==3.0.48
protobuf==4.24.4
pyasn1==0.5.1
pydantic==1.8.2
Pygments==2.17.2
pyparsing==3.1.4
pytest==7.4.4
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-Levenshtein==0.12.1
pytorch-pretrained-bert==0.6.2
pytz==2025.2
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
responses==0.23.3
rsa==4.7.2
s3transfer==0.8.2
sacremoses==0.0.53
scikit-learn==0.20.0
scipy==1.7.3
sentencepiece==0.1.95
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
spacy==2.1.9
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.4
srsly==1.0.7
tensorboardX==2.6.2.2
text-unidecode==1.3
thinc==7.0.8
tokenizers==0.10.3
tomli==2.0.1
torch==1.13.1
tqdm==4.67.1
transformers==4.11.3
types-PyYAML==6.0.12.12
typing-utils==0.1.0
typing_extensions==4.7.1
Unidecode==1.4.0
urllib3==1.26.20
uvicorn==0.21.1
uvloop==0.18.0
vine==5.1.0
wasabi==0.10.1
watchfiles==0.20.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==2.2.3
word2number==1.1
zipp==3.15.0
zope.event==5.0
zope.interface==6.4.post2
